+++ config/configunmarshaler/defaultunmarshaler_test.go
@@ -15,7 +15,6 @@
 package configunmarshaler
 
 import (
-	"context"
 	"path/filepath"
 	"testing"
 
@@ -26,9 +25,9 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/configmapprovider"
 	"go.opentelemetry.io/collector/config/confignet"
 	"go.opentelemetry.io/collector/config/configtelemetry"
+	"go.opentelemetry.io/collector/config/configtest"
 	"go.opentelemetry.io/collector/internal/testcomponents"
 )
 
@@ -227,9 +226,7 @@ func TestLoadEmptyAllSections(t *testing.T) {
 }
 
 func loadConfigFile(t *testing.T, fileName string, factories component.Factories) (*config.Config, error) {
-	v, err := configmapprovider.NewFile(fileName).Retrieve(context.Background(), nil)
-	require.NoError(t, err)
-	cm, err := v.Get(context.Background())
+	cm, err := configtest.LoadConfigMap(fileName)
 	require.NoError(t, err)
 
 	// Unmarshal the config from the config.Map using the given factories.